+++ install.php
@@ -1,13 +1,17 @@
 <?php
 
-/* @license   <a href="http://www.gnu.org/licenses/gpl-3.0.html" target="_blank">GNU/GPLv3</a> */
+/* @license   <a href="https://www.gnu.org/licenses/gpl-3.0.html" target="_blank">GNU/GPLv3</a> */
 
 // doc: https://docs.joomla.org/J3.x:Creating_a_simple_module/Adding_an_install-uninstall-update_script_file/fr
-// voir flexicontactplus
+
 // No direct access
 defined('_JEXEC') or die('Restricted access');
 
 use Joomla\CMS\Factory;
+use Joomla\CMS\Version;
+use Joomla\Filesystem\File;
+use Joomla\Filesystem\Folder;
+use Joomla\Database\DatabaseInterface;
 
 /*
  *  Première installation de UP :
@@ -25,6 +29,9 @@ class plgContentUpInstallerScript {
     
 	private $dir  = null;
 	private $lang = null;
+	private $min_joomla_version      = '5.2.0';
+	private $min_php_version         = '8.1';
+    private $installerName = 'plgcontentupinstaller';
 
 	public function __construct()
 	{
@@ -61,6 +68,19 @@ class plgContentUpInstallerScript {
      * @return void
      */
     public function preflight($type, $parent) {
+        
+		if ( ! $this->passMinimumJoomlaVersion())
+		{
+			$this->uninstallInstaller();
+			return false;
+		}
+
+		if ( ! $this->passMinimumPHPVersion())
+		{
+			$this->uninstallInstaller();
+			return false;
+		}
+        
         $app = Factory::getApplication();
         // $app->enqueueMessage('<p>actions avant l\'installation/mise à jour/désinstallation du plugin</p>');
         $path = JPATH_ROOT . '/plugins/content/up/';
@@ -175,5 +195,75 @@ class plgContentUpInstallerScript {
         } 
         return true;
     }
+	// Check if Joomla version passes minimum requirement
+	private function passMinimumJoomlaVersion()
+	{
+		$j = new Version();
+		$version=$j->getShortVersion(); 
+		if (version_compare($version, $this->min_joomla_version, '<'))
+		{
+			Factory::getApplication()->enqueueMessage(
+				'Incompatible Joomla version : found <strong>' . $version . '</strong>, Minimum : <strong>' . $this->min_joomla_version . '</strong>',
+				'error'
+			);
 
+			return false;
+		}
+
+		return true;
+	}
+
+	// Check if PHP version passes minimum requirement
+	private function passMinimumPHPVersion()
+	{
+
+		if (version_compare(PHP_VERSION, $this->min_php_version, '<'))
+		{
+			Factory::getApplication()->enqueueMessage(
+					'Incompatible PHP version : found  <strong>' . PHP_VERSION . '</strong>, Minimum <strong>' . $this->min_php_version . '</strong>',
+				'error'
+			);
+			return false;
+		}
+
+		return true;
+	}
+	private function uninstallInstaller()
+	{
+		if ( ! is_dir::exists(JPATH_PLUGINS . '/system/' . $this->installerName)) {
+			return;
+		}
+		$this->delete([
+			JPATH_PLUGINS . '/system/' . $this->installerName . '/language',
+			JPATH_PLUGINS . '/system/' . $this->installerName,
+		]);
+		$db = Factory::getContainer()->get(DatabaseInterface::class);
+		$query = $db->createQuery()
+			->delete('#__extensions')
+			->where($db->quoteName('element') . ' = ' . $db->quote($this->installerName))
+			->where($db->quoteName('folder') . ' = ' . $db->quote('system'))
+			->where($db->quoteName('type') . ' = ' . $db->quote('plugin'));
+		$db->setQuery($query);
+		$db->execute();
+        $cacheModel = Factory::getApplication()->bootComponent('com_cache')->getMVCFactory()->createModel('Cache', 'Administrator', ['ignore_request' => true]);
+        $cache = $cacheModel->getCache() ??null;
+        if ($cache) {
+            foreach ($cache->getAll() as $group) {
+                $cache->clean($group->group);
+            }
+            Factory::getApplication()->enqueueMessage('<p>Cache Ok.</p>');
+        }
+	}
+    public function delete($files = [])
+    {
+        foreach ($files as $file) {
+            if (is_dir($file)) {
+                Folder::delete($file);
+            }
+
+            if (is_file($file)) {
+                File::delete($file);
+            }
+        }
+    }
 }