+++ changelog.md
@@ -5,6 +5,7 @@
 - UP devient un "service provider"
 - déplacement des fonctions dans UpHelper
 - scssphp version 2.1.0 + uri 7.7.0
+- install : vérifie la version minimale de Joomla et PHP 
 
 ## 06/12/2025 - version 5.4.12